Update README.md
I added a line in the curtain section of the documentation to add the pause command. This command is inside the API itself, but it is undocumented. I imagine this is unintentional, likely an over sight, and should be fixed, because it does work. I tested in v1.1 of the API no issue with command:pause.