Update the days until stale, and days until close values of the GitHub Stale Bot.
The users of this repo are often very frustrated, when their carefully opened issues are marked stale after 10 days and then closed.
In addition, it can make SwitchBot customers get the very strong impression that the API is not maintained,
or that the API maintainers are completely disinterested in open bugs and feature requests.
Further, it causes false activity just for the sake of keeping issues open, which makes for unnecessary traffic.

Almost no issues seem to be addressed within the 10 days that were configured.

This PR configures the 10 days, to a much more reasonable two years.
Furthermore, it gives five days instead of one day, between the staleness warning and the issue close.
For people not using GitHub every day, it is hard to notice a notification in time after just one day.
In addition, even for those with email notifications active, checking email every day is becoming less common.